Don't attempt to install/upgrade packages with brew if it will fail
In this case, the user is a secondary non-admin user on a mac that did
not initially setup/install homebrew. Instead of waiting to return from
`do_install` after `brew update` fails (with a message from brew telling
the user to try to `sudo chown -R` /usr/local/Cellar), return from
`do_install` immediately if /usr/local/Cellar is not writeable by the
user that called it.